Reorganize integration test folder
