+++ lib/src/__tests__/Theme.test.tsx
@@ -1,9 +1,9 @@
 import React from 'react';
-import { mount } from '../test-utils';
-import { DatePicker } from '../../DatePicker';
+import { mount } from './test-utils';
+import { DatePicker } from '../DatePicker';
 import { ThemeProvider } from '@material-ui/core/styles';
 import { createMuiTheme } from '@material-ui/core/styles';
-import { DateTimePicker } from '../../DateTimePicker/DateTimePicker';
+import { DateTimePicker } from '../DateTimePicker/DateTimePicker';
 
 const theme = createMuiTheme({
   palette: {
