+++ lib/src/__tests__/commands.tsx
