+++ lib/src/__tests__/test-utils.tsx
@@ -5,8 +5,8 @@ import MomentUtils from '@date-io/moment';
 import DateFnsUtils from '@date-io/date-fns';
 import LocalizationProvider from '../LocalizationProvider';
 import { IUtils } from '@date-io/core/IUtils';
-import { createMuiTheme, ThemeProvider } from '@material-ui/core/styles';
 import { MaterialUiPickersDate } from '../typings/date';
+import { createMuiTheme, ThemeProvider } from '@material-ui/core/styles';
 
 interface WithUtilsProps {
   utils: IUtils<MaterialUiPickersDate>;
