+++ lib/src/__tests__/unit/date-utils.test.tsx
@@ -1,5 +1,5 @@
-import { utilsToUse } from './test-utils';
-import { findClosestEnabledDate } from '../_helpers/date-utils';
+import { utilsToUse } from '../test-utils';
+import { findClosestEnabledDate } from '../../_helpers/date-utils';
 
 describe('findClosestEnabledDate', () => {
   const day18thText = utilsToUse.format(utilsToUse.date('2018-08-18'), 'dayOfMonth');
